feat: Refactor chat components and add new model selection and question block features
