+++ components/private/chat/ChatClient.tsx
@@ -12,8 +12,8 @@ import ChatInput from "./ChatInput";
 import ChatSidePanel from "@/components/private/chat/ChatSidePanel";
 
 // PMP-specific components
-import QuestionBlock from "@/components/private/PMP/QuestionBlock";
-import ModelSelection from "@/components/private/PMP/ModelSelection";
+import QuestionBlock from "@/components/private/modes/PMP/QuestionBlock";
+import ModelSelection from "@/components/private/modes/PMP/ModelSelection";
 
 interface ChatClientProps {
   user: User;
