+++ components/private/modes/Compare/CompareResponses.tsx
